Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib
pyjwt